refactoring autocompletes
